foncorp/cf_xa_dcx/actualizar-viaje-tool/requirements.txt
functions-framework>=3.0.0
Flask>=2.0.0
google-cloud-bigquery>=3.0.0